+++ src/BookPage.php
@@ -47,7 +47,7 @@ class BookPage {
   /**
    * @param mixed $parameter
    */
-  public function addCondition(IBookPageCondition $condition, $parameter): void {
+  public function addCondition(BookPageCondition $condition, $parameter): void {
     $this->conditions[] = [$condition, $parameter];
   }