+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 2.4.0-dev
 - raised minimal version of PHP to 8.1
+- deprecated IBookPageCondition in favor of new BookPageCondition
 
 Version 2.3.0
 - raised minimal version of PHP to 7.4